fix: enforce case-sensitive renames in git for components
