+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "a4chat",
-  "version": "3.0.0",
+  "version": "3.1.0",
   "private": true,
   "scripts": {
     "postinstall": "node scripts/prisma-sync.mjs",
