+++ src-tauri/tauri.conf.json
@@ -1,7 +1,7 @@
 {
   "$schema": "https://schema.tauri.app/config/2",
   "productName": "A4Chat",
-  "version": "3.0.0",
+  "version": "3.1.0",
   "identifier": "com.ariajsarkar.a4chat",
   "build": {
     "beforeDevCommand": "node scripts/dev-server.mjs",
