+++ src/components/Settings/Pairing.tsx
