+++ src/lib/AppMeta.ts
@@ -1,4 +1,4 @@
 export const APP_NAME = "A4Chat";
-export const APP_VERSION = "3.0.0";
+export const APP_VERSION = "3.1.0";
 export const GITHUB_URL = "https://github.com/AriajSarkar/A4Chat";
 export const GITHUB_RELEASES_URL = `${GITHUB_URL}/releases`;
